requirements.txt
Flask==0.12.2
line-bot-sdk==1.5.0
pymongo==3.7.2
gunicorn==19.9.0
